Run the N-sweep: the decomposition holds
30 ensembles, tail split. Injected noise recovered at 0.0208
against a true 0.020 while misfit fell from 0.0398 to 0.0111,
so the reducible and irreducible halves separate as claimed.
Epistemic shrinks monotonically in both conditions.

Also: the extrapolation gap widens with data, 1.71 to 3.39.
A drafted claim that out-of-region calibration degrades with N
was withdrawn, it was one seed and is flat across three.

nested_subsamples keeps the rungs nested, results checkpoint
after every cell and resume on restart.

+++ scripts/deferral.py
@@ -79,7 +79,7 @@ def curve(crps, signal, rates):
     cumulative = np.concatenate([[0.0], np.cumsum(np.asarray(crps)[order])])
     n = len(crps)
     total = cumulative[-1]
-    return np.array([(total - cumulative[int(round(r * n))]) / n for r in rates])
+    return np.array([(total - cumulative[round(r * n)]) / n for r in rates])
 
 
 def random_curve(crps, rates, seed):
@@ -157,7 +157,7 @@ def main():
         subset = df[df['region'] == region]
         curves, summary = run_region(subset)
 
-        payload[region] = {'n': int(len(subset)), 'rankings': summary}
+        payload[region] = {'n': len(subset), 'rankings': summary}
         print_region(f'{label}, n = {len(subset):,}', summary)
 
         for name, values in curves.items():
